+++ mgcpy/independence_tests/utils/compute_distance_matrix.py
@@ -2,6 +2,7 @@
     **Common Distance Calculation Matrix**
 """
 
+
 def compute_distance(matrix_X, matrix_Y, _compute_distance):
     """
     Computes the distance matrix among both independence tests
@@ -24,7 +25,7 @@ def compute_distance(matrix_X, matrix_Y, _compute_distance):
     # use the matrix shape and diagonal elements to determine if the given data is a distance matrix or not
     if matrix_X.shape[0] != matrix_X.shape[1] or sum(matrix_X.diagonal()**2) > 0:
         matrix_X = _compute_distance(matrix_X)
-    if matrix_Y.shape[0] != matrix_Y.shape[1] or sum(matrix_Y.diagonal()**2) > 0:
+    if matrix_X.shape[0] != matrix_X.shape[1] or sum(matrix_Y.diagonal()**2) > 0:
         matrix_Y = _compute_distance(matrix_Y)
 
     return matrix_X, matrix_Y