+++ CHANGELOG.md
@@ -3,8 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.2.1] - 2019-02-07
+## [0.3.0] - 2019-02-07
+- Refactor the mgcpy package structure
+- Add Manova
 - Fix two sample tests
+- Add scripts for k sample test power curves
 - Update sphinx docs
 
 ## [0.2.0] - 2019-01-25